+++ hazelcast/include/hazelcast/client/spi/InvocationService.h
@@ -23,8 +23,10 @@
 #include "hazelcast/util/HazelcastDll.h"
 #include "hazelcast/util/AtomicInt.h"
 #include "hazelcast/util/SynchronizedMap.h"
-#include <boost/shared_ptr.hpp>
 #include "hazelcast/client/protocol/IMessageHandler.h"
+
+#include <boost/shared_ptr.hpp>
+#include <stdint.h>
 
 #if  defined(WIN32) || defined(_WIN32) || defined(WIN64) || defined(_WIN64)
 #pragma warning(push)
@@ -101,7 +103,7 @@ namespace hazelcast {
                 * @param callId of event handler registration request
                 * @return true if found and removed, false otherwise
                 */
-                void removeEventHandler(int callId);
+                void removeEventHandler(int64_t callId);
 
                 /**
                 * Clean all promises (both request and event handlers). Retries requests on available connections if applicable.
@@ -128,8 +130,8 @@ namespace hazelcast {
                 int retryWaitTime;
                 int retryCount;
                 spi::ClientContext& clientContext;
-                util::SynchronizedMap<connection::Connection* , util::SynchronizedMap<int, connection::CallPromise > > callPromises;
-                util::SynchronizedMap<connection::Connection* , util::SynchronizedMap<int, connection::CallPromise > > eventHandlerPromises;
+                util::SynchronizedMap<connection::Connection* , util::SynchronizedMap<int64_t, connection::CallPromise > > callPromises;
+                util::SynchronizedMap<connection::Connection* , util::SynchronizedMap<int64_t, connection::CallPromise > > eventHandlerPromises;
 
                 bool isAllowedToSentRequest(connection::Connection& connection, protocol::ClientMessage const&);
 
@@ -145,13 +147,13 @@ namespace hazelcast {
 
                 void registerCall(connection::Connection& connection, boost::shared_ptr<connection::CallPromise> promise);
 
-                boost::shared_ptr<connection::CallPromise> deRegisterCall(connection::Connection& connection, int callId);
+                boost::shared_ptr<connection::CallPromise> deRegisterCall(connection::Connection& connection, int64_t callId);
 
                 /** **/
-                void registerEventHandler(int correlationId,
+                void registerEventHandler(int64_t correlationId,
                                           connection::Connection& connection, boost::shared_ptr<connection::CallPromise> promise);
 
-                boost::shared_ptr<connection::CallPromise> deRegisterEventHandler(connection::Connection& connection, int callId);
+                boost::shared_ptr<connection::CallPromise> deRegisterEventHandler(connection::Connection& connection, int64_t callId);
 
                 /***** HANDLE PACKET PART ****/
 
@@ -163,15 +165,15 @@ namespace hazelcast {
 
                 /** CallPromise Map **/
 
-                boost::shared_ptr< util::SynchronizedMap<int, connection::CallPromise> > getCallPromiseMap(connection::Connection& connection);
+                boost::shared_ptr< util::SynchronizedMap<int64_t, connection::CallPromise> > getCallPromiseMap(connection::Connection& connection);
 
                 /** EventHandler Map **/
 
                 // TODO: Put the promise map as a member of the connection object. In this way, we can get the promise map directly from connection object
                 // without a need for a map lookup since we already know the connection and the map is specific to a connection
-                boost::shared_ptr< util::SynchronizedMap<int, connection::CallPromise> > getEventHandlerPromiseMap(connection::Connection& connection);
+                boost::shared_ptr< util::SynchronizedMap<int64_t, connection::CallPromise> > getEventHandlerPromiseMap(connection::Connection& connection);
 
-                boost::shared_ptr<connection::CallPromise> getEventHandlerPromise(connection::Connection& , int callId);
+                boost::shared_ptr<connection::CallPromise> getEventHandlerPromise(connection::Connection& , int64_t callId);
             };
         }
     }
